+++ extensions/distill/wrapper-validation.test.ts
@@ -34,168 +34,20 @@ import * as fs from "node:fs";
 import * as os from "node:os";
 import * as path from "node:path";
 
-import { SessionManager } from "@earendil-works/pi-coding-agent";
-
-import { withNapkinOnPath } from "./_test-helpers";
-import { createDistillWorkspace } from "./distill-workspace";
+import {
+  makeWrapperScaffold,
+  runWrapperWithStub,
+  withNapkinOnPath,
+  writePiStub,
+} from "./_test-helpers";
 import { DISTILL_WRAPPER_SCRIPT } from "./scripts-paths";
 
-interface Scaffold {
-  root: string;
-  vault: string;
-  parentCwd: string;
-  sessionFile: string;
-  errorDir: string;
-  stubPi: string;
-}
-
-/**
- * Build a fresh test scaffold per test. Creates a git-init'd vault
- * with one seed commit (so `<seed-sha>..HEAD` rev-list semantics work),
- * a parent cwd, an empty session file, and the error dir. Caller must
- * call `cleanup` in afterEach.
- */
-function makeScaffold(): Scaffold {
-  const root = fs.mkdtempSync(path.join(os.tmpdir(), "napkin-distill-a3-"));
-  const vault = path.join(root, "vault");
-  const parentCwd = path.join(root, "parent");
-  const sessionsDir = path.join(root, "sessions");
-  const errorDir = path.join(vault, ".napkin", "distill", "errors");
-  const stubPi = path.join(root, "stub-pi");
-
-  fs.mkdirSync(vault);
-  fs.mkdirSync(parentCwd);
-  fs.mkdirSync(sessionsDir);
-  fs.mkdirSync(errorDir, { recursive: true });
-
-  // git init + seed commit. Use -b main so detectDefaultBranch resolves.
-  spawnSync("git", ["init", "-b", "main", vault], { encoding: "utf-8" });
-  spawnSync("git", ["-C", vault, "config", "user.email", "test@example.com"]);
-  spawnSync("git", ["-C", vault, "config", "user.name", "test"]);
-  fs.writeFileSync(path.join(vault, "seed.md"), "# seed\n");
-  spawnSync("git", ["-C", vault, "add", "."]);
-  spawnSync("git", ["-C", vault, "commit", "-m", "seed"]);
-
-  const sessionFile = (() => {
-    const sm = SessionManager.create(parentCwd, sessionsDir);
-    sm.appendMessage({ role: "user", content: "hello" });
-    sm.appendMessage({ role: "assistant", content: "hi" });
-    const file = sm.getSessionFile();
-    if (!file || !fs.existsSync(file)) {
-      throw new Error("failed to create test session on disk");
-    }
-    return file;
-  })();
-
-  return { root, vault, parentCwd, sessionFile, errorDir, stubPi };
-}
-
-/**
- * Write a stub `pi` binary. The body is whatever the test wants the
- * agent to do; positional args are ignored unless the body parses them.
- * `chmod +x`, then return the path so the caller can pass it as
- * `NAPKIN_DISTILL_PI_BIN`.
- */
-function writeStubPi(scaffold: Scaffold, bodyScript: string): string {
-  const stub = `#!/usr/bin/env bash\nset -e\n${bodyScript}\n`;
-  fs.writeFileSync(scaffold.stubPi, stub, { mode: 0o755 });
-  return scaffold.stubPi;
-}
-
-/**
- * Run the wrapper end-to-end and return its exit code, stderr, and the
- * outcome sidecar path (if any was written).
- *
- * The wrapper's argv shape (PR #12 A2): vault, worktree, branch,
- * sessionFork, prompt, errorDir, model, defaultBranch, parentCwd,
- * maxDurationSecs.
- */
-function runWrapper(
-  scaffold: Scaffold,
-  opts: {
-    skipPi?: boolean;
-    extraEnv?: Record<string, string>;
-    maxDurationSecs?: string;
-  } = {},
-): {
-  exitCode: number;
-  stderr: string;
-  outcome: string | null;
-  outcomePath: string | null;
-  branch: string;
-  workspace: ReturnType<typeof createDistillWorkspace>;
-} {
-  const workspace = createDistillWorkspace(
-    scaffold.vault,
-    scaffold.sessionFile,
-    scaffold.parentCwd,
-  );
-  const branch = workspace.branchName;
-
-  const env: Record<string, string> = {
-    ...process.env,
-    GIT_AUTHOR_NAME: "test",
-    GIT_AUTHOR_EMAIL: "test@example.com",
-    GIT_COMMITTER_NAME: "test",
-    GIT_COMMITTER_EMAIL: "test@example.com",
-    NAPKIN_DISTILL_NO_RECURSE: "1",
-    NAPKIN_DISTILL_PI_BIN: scaffold.stubPi,
-    ...(opts.extraEnv ?? {}),
-  };
-  if (opts.skipPi) {
-    env.NAPKIN_DISTILL_SKIP_PI = "1";
-  }
-
-  const r = spawnSync(
-    "bash",
-    [
-      DISTILL_WRAPPER_SCRIPT,
-      scaffold.vault,
-      workspace.worktreePath,
-      branch,
-      workspace.sessionForkPath,
-      "test prompt",
-      scaffold.errorDir,
-      "",
-      "main",
-      scaffold.parentCwd,
-      opts.maxDurationSecs ?? "60",
-    ],
-    {
-      cwd: scaffold.parentCwd,
-      encoding: "utf-8",
-      env,
-    },
-  );
-
-  // Locate the outcome sidecar. The wrapper names it
-  // `<ts>-<pid>-<branchShort>.outcome`. PR #12 A4 made the file
-  // multi-line for `failed:*` classes (line 1 = class, lines 2+ =
-  // recovery hint); use only line 1 as the canonical class string
-  // — same shape as the JS-side `findDistillOutcomeForBranch`.
-  const branchShort = branch.replace(/^distill\//, "");
-  const outcomeFiles = fs.existsSync(scaffold.errorDir)
-    ? fs
-        .readdirSync(scaffold.errorDir)
-        .filter((f) => f.endsWith(`-${branchShort}.outcome`))
-    : [];
-  let outcome: string | null = null;
-  let outcomePath: string | null = null;
-  if (outcomeFiles.length === 1) {
-    outcomePath = path.join(scaffold.errorDir, outcomeFiles[0]);
-    const raw = fs.readFileSync(outcomePath, "utf-8");
-    outcome = (raw.split("\n")[0] ?? "").trim();
-  }
-
-  return {
-    exitCode: r.status ?? -1,
-    stderr: r.stderr ?? "",
-    outcome,
-    outcomePath,
-    branch,
-    workspace,
-  };
-}
+// Local aliases keep test bodies aligned with the prior in-file helpers
+// so the CLEAN-A-6 extraction is a pure call-site move with no per-test
+// edits. Phase C may switch new tests to the canonical helper names.
+const makeScaffold = () => makeWrapperScaffold("napkin-distill-a3-");
+const writeStubPi = writePiStub;
+const runWrapper = runWrapperWithStub;
 
 describe("distill-wrapper.sh post-agent validation (PR #12 A3)", () => {
   let pathHandle: { restore: () => void };
